sitemap

+++ src/api/achievements/id/comment/mod.rs
@@ -42,7 +42,7 @@ pub fn configure(cfg: &mut web::ServiceConfig) {
 #[put("/api/achievements/{id}/comment")]
 async fn put_achievement_comment(
     session: Session,
-    id: web::Path<i64>,
+    id: web::Path<i32>,
     comment_update: web::Json<CommentUpdate>,
     pool: web::Data<PgPool>,
 ) -> ApiResult<impl Responder> {
@@ -75,7 +75,7 @@ async fn put_achievement_comment(
 #[delete("/api/achievements/{id}/comment")]
 async fn delete_achievement_comment(
     session: Session,
-    id: web::Path<i64>,
+    id: web::Path<i32>,
     pool: web::Data<PgPool>,
 ) -> ApiResult<impl Responder> {
     let Ok(Some(username)) = session.get::<String>("username") else {
