+++ src/api/books/id/comment/mod.rs
@@ -41,7 +41,7 @@ pub fn configure(cfg: &mut web::ServiceConfig) {
 #[put("/api/books/{id}/comment")]
 async fn put_book_comment(
     session: Session,
-    id: web::Path<i64>,
+    id: web::Path<i32>,
     comment_update: web::Json<CommentUpdate>,
     pool: web::Data<PgPool>,
 ) -> ApiResult<impl Responder> {
@@ -74,7 +74,7 @@ async fn put_book_comment(
 #[delete("/api/books/{id}/comment")]
 async fn delete_book_comment(
     session: Session,
-    id: web::Path<i64>,
+    id: web::Path<i32>,
     pool: web::Data<PgPool>,
 ) -> ApiResult<impl Responder> {
     let Ok(Some(username)) = session.get::<String>("username") else {
