+++ src/database/achievements.rs
@@ -3,7 +3,7 @@ use sqlx::PgPool;
 
 #[derive(Clone)]
 pub struct DbAchievement {
-    pub id: i64,
+    pub id: i32,
     pub series: i32,
     pub series_name: String,
     pub name: String,
@@ -109,7 +109,7 @@ pub async fn get_achievements(language: &str, pool: &PgPool) -> Result<Vec<DbAch
     .await?)
 }
 
-pub async fn get_achievements_id(pool: &PgPool) -> Result<Vec<i64>> {
+pub async fn get_achievements_id(pool: &PgPool) -> Result<Vec<i32>> {
     Ok(sqlx::query!(
         "
         SELECT
@@ -127,7 +127,7 @@ pub async fn get_achievements_id(pool: &PgPool) -> Result<Vec<i64>> {
     .collect())
 }
 
-pub async fn get_related(id: i64, set: i32, pool: &PgPool) -> Result<Vec<i64>> {
+pub async fn get_related(id: i32, set: i32, pool: &PgPool) -> Result<Vec<i32>> {
     Ok(sqlx::query!(
         "
         SELECT
@@ -150,7 +150,7 @@ pub async fn get_related(id: i64, set: i32, pool: &PgPool) -> Result<Vec<i64>> {
 }
 
 pub async fn get_achievement_by_id(
-    id: i64,
+    id: i32,
     language: &str,
     pool: &PgPool,
 ) -> Result<DbAchievement> {
@@ -188,7 +188,7 @@ pub async fn get_achievement_by_id(
     .await?)
 }
 
-pub async fn update_achievement_version(id: i64, version: &str, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_version(id: i32, version: &str, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET version = $2 WHERE id = $1",
         id,
@@ -200,7 +200,7 @@ pub async fn update_achievement_version(id: i64, version: &str, pool: &PgPool) -
     Ok(())
 }
 
-pub async fn update_achievement_comment(id: i64, comment: &str, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_comment(id: i32, comment: &str, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET comment = $2 WHERE id = $1",
         id,
@@ -212,7 +212,7 @@ pub async fn update_achievement_comment(id: i64, comment: &str, pool: &PgPool) -
     Ok(())
 }
 
-pub async fn update_achievement_reference(id: i64, reference: &str, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_reference(id: i32, reference: &str, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET reference = $2 WHERE id = $1",
         id,
@@ -224,7 +224,7 @@ pub async fn update_achievement_reference(id: i64, reference: &str, pool: &PgPoo
     Ok(())
 }
 
-pub async fn update_achievement_difficulty(id: i64, difficulty: &str, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_difficulty(id: i32, difficulty: &str, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET difficulty = $2 WHERE id = $1",
         id,
@@ -236,7 +236,7 @@ pub async fn update_achievement_difficulty(id: i64, difficulty: &str, pool: &PgP
     Ok(())
 }
 
-pub async fn update_achievement_gacha(id: i64, gacha: bool, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_gacha(id: i32, gacha: bool, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET gacha = $2 WHERE id = $1",
         id,
@@ -248,7 +248,7 @@ pub async fn update_achievement_gacha(id: i64, gacha: bool, pool: &PgPool) -> Re
     Ok(())
 }
 
-pub async fn update_achievement_impossible(id: i64, impossible: bool, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_impossible(id: i32, impossible: bool, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET impossible = $2 WHERE id = $1",
         id,
@@ -260,7 +260,7 @@ pub async fn update_achievement_impossible(id: i64, impossible: bool, pool: &PgP
     Ok(())
 }
 
-pub async fn update_achievement_video(id: i64, video: &str, pool: &PgPool) -> Result<()> {
+pub async fn update_achievement_video(id: i32, video: &str, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET video = $2 WHERE id = $1",
         id,
@@ -272,7 +272,7 @@ pub async fn update_achievement_video(id: i64, video: &str, pool: &PgPool) -> Re
     Ok(())
 }
 
-pub async fn delete_achievement_version(id: i64, pool: &PgPool) -> Result<()> {
+pub async fn delete_achievement_version(id: i32, pool: &PgPool) -> Result<()> {
     sqlx::query!("UPDATE achievements SET version = NULL WHERE id = $1", id,)
         .execute(pool)
         .await?;
@@ -280,7 +280,7 @@ pub async fn delete_achievement_version(id: i64, pool: &PgPool) -> Result<()> {
     Ok(())
 }
 
-pub async fn delete_achievement_comment(id: i64, pool: &PgPool) -> Result<()> {
+pub async fn delete_achievement_comment(id: i32, pool: &PgPool) -> Result<()> {
     sqlx::query!("UPDATE achievements SET comment = NULL WHERE id = $1", id)
         .execute(pool)
         .await?;
@@ -288,7 +288,7 @@ pub async fn delete_achievement_comment(id: i64, pool: &PgPool) -> Result<()> {
     Ok(())
 }
 
-pub async fn delete_achievement_reference(id: i64, pool: &PgPool) -> Result<()> {
+pub async fn delete_achievement_reference(id: i32, pool: &PgPool) -> Result<()> {
     sqlx::query!("UPDATE achievements SET reference = NULL WHERE id = $1", id)
         .execute(pool)
         .await?;
@@ -296,7 +296,7 @@ pub async fn delete_achievement_reference(id: i64, pool: &PgPool) -> Result<()> 
     Ok(())
 }
 
-pub async fn delete_achievement_difficulty(id: i64, pool: &PgPool) -> Result<()> {
+pub async fn delete_achievement_difficulty(id: i32, pool: &PgPool) -> Result<()> {
     sqlx::query!(
         "UPDATE achievements SET difficulty = NULL WHERE id = $1",
         id,
@@ -307,7 +307,7 @@ pub async fn delete_achievement_difficulty(id: i64, pool: &PgPool) -> Result<()>
     Ok(())
 }
 
-pub async fn delete_achievement_video(id: i64, pool: &PgPool) -> Result<()> {
+pub async fn delete_achievement_video(id: i32, pool: &PgPool) -> Result<()> {
     sqlx::query!("UPDATE achievements SET video = NULL WHERE id = $1", id)
         .execute(pool)
         .await?;
